# doc/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "华农掌中行 API",
    "version": "1.0.0",
    "description": "由 back/server.js 与 back/routes/*.js 导出的接口定义（2026-08-14）。"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "本地后端（以 PORT 环境变量为准）"
    }
  ],
...
  "paths": {
...
    "/api/charging/stations": {
      "get": {
        "tags": [
          "充电"
        ],
        "summary": "充电桩列表",
        "operationId": "get_api_charging_stations",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "400": {
            "description": "请求参数或业务状态错误"
          },
          "401": {
            "description": "未认证或 token 无效"
          },
          "403": {
            "description": "权限不足"
          },
          "404": {
            "description": "资源不存在"
          },
          "500": {
            "description": "服务器错误"
          }
        },
        "parameters": [
          {
            "name": "latitude",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "longitude",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number"
            }
          }
        ]
      }
    },
    "/api/charging/reserve": {
      "post": {
        "tags": [
          "充电"
        ],
        "summary": "预约充电桩",
        "description": "userId 也可由 userid 请求头传入；未传时当前实现默认为 1",
        "operationId": "post_api_charging_reserve",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "400": {
            "description": "请求参数或业务状态错误"
          },
          "401": {
            "description": "未认证或 token 无效"
          },
          "403": {
            "description": "权限不足"
          },
          "404": {
            "description": "资源不存在"
          },
          "500": {
            "description": "服务器错误"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "stationId": {
                    "type": "number"
                  },
                  "startTime": {
                    "type": "string"
                  },
                  "duration": {
                    "type": "number",
                    "default": 60
                  },
                  "phone": {
                    "type": "string"
                  },
                  "userId": {
                    "type": "number"
                  }
                },
                "required": [
                  "stationId",
                  "startTime"
                ]
              }
            }
          }
        }
      }
    },
    "/api/charging/station/{id}": {
      "get": {
        "tags": [
          "充电"
        ],
        "summary": "充电桩详情",
        "operationId": "get_api_charging_station_id",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "400": {
            "description": "请求参数或业务状态错误"
          },
          "401": {
            "description": "未认证或 token 无效"
          },
          "403": {
            "description": "权限不足"
          },
          "404": {
            "description": "资源不存在"
          },
          "500": {
            "description": "服务器错误"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "number"
            }
          }
        ]
      }
    },
    "/api/charging/cancel/{id}": {
      "post": {
        "tags": [
          "充电"
        ],
        "summary": "取消预约",
        "operationId": "post_api_charging_cancel_id",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "400": {
            "description": "请求参数或业务状态错误"
          },
          "401": {
            "description": "未认证或 token 无效"
          },
          "403": {
            "description": "权限不足"
          },
          "404": {
            "description": "资源不存在"
          },
          "500": {
            "description": "服务器错误"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "number"
            }
          }
        ]
      }
    },
    "/api/charging/records": {
      "get": {
        "tags": [
          "充电"
        ],
        "summary": "用户充电记录",
        "operationId": "get_api_charging_records",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "400": {
            "description": "请求参数或业务状态错误"
          },
          "401": {
            "description": "未认证或 token 无效"
          },
          "403": {
            "description": "权限不足"
          },
          "404": {
            "description": "资源不存在"
          },
          "500": {
            "description": "服务器错误"
          }
        },
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number"
            }
          }
        ]
      }
    },
    "/api/charging/reservation": {
      "get": {
        "tags": [
          "充电"
        ],
        "summary": "用户当前预约",
        "operationId": "get_api_charging_reservation",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "400": {
            "description": "请求参数或业务状态错误"
          },
          "401": {
            "description": "未认证或 token 无效"
          },
          "403": {
            "description": "权限不足"
          },
          "404": {
            "description": "资源不存在"
          },
          "500": {
            "description": "服务器错误"
          }
        },
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "stationId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number"
            }
          }
        ]
      }
    },
    "/api/charging/reservations": {
      "get": {
        "tags": [
          "充电"
        ],
        "summary": "用户全部预约",
        "operationId": "get_api_charging_reservations",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "400": {
            "description": "请求参数或业务状态错误"
          },
          "401": {
            "description": "未认证或 token 无效"
          },
          "403": {
            "description": "权限不足"
          },
          "404": {
            "description": "资源不存在"
          },
          "500": {
            "description": "服务器错误"
          }
        },
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number"
            }
          }
        ]
      }
    },
...
    "/api/charging/suggestions": {
      "get": {
        "tags": [
          "充电"
        ],
        "summary": "用户规划建议列表",
        "operationId": "get_api_charging_suggestions",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "400": {
            "description": "请求参数或业务状态错误"
          },
          "401": {
            "description": "未认证或 token 无效"
          },
          "403": {
            "description": "权限不足"
          },
          "404": {
            "description": "资源不存在"
          },
          "500": {
            "description": "服务器错误"
          }
        },
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number"
            }
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "userToken": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "用户 JWT，直接传 token"
      },
      "adminBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "管理员 JWT"
      }
    },
    "schemas": {
      "ApiResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "example": 0
          },
          "message": {
            "type": "string"
          },
          "data": {
            "nullable": true
          }
        },
        "required": [
          "code"
        ]
      }
    }
  }
}